BOL-375 refactor: move Auth Datasource's package
